requirements.txt
beautifulsoup4==4.9.1 
kivy==2.0.0
requests==2.25.1 
pillow==8.2.0
pyzbar==0.1.8
opencv-python==4.5.1.48
